refactor: remove @zeltjs/kv-driver-redis package
Migrate Redis KV functionality to @zeltjs/kv/adaptor-redis subpath.
Update @zeltjs/testing to use @zeltjs/redis directly.
Fix adaptor file naming to use .adaptor.ts suffix.

Verified: precommit-verify ✓ (4aa7faa8863e877c)
