--- packages/kv-driver-redis/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@zeltjs/kv-driver-redis",
-  "version": "0.1.0",
-  "type": "module",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zeltjs/zelt.git",
-    "directory": "packages/kv-driver-redis"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsdown",
-    "test": "vitest run",
-    "typecheck": "tsc -b"
-  },
-  "peerDependencies": {
-    "@zeltjs/core": "workspace:*",
-    "@zeltjs/kv": "workspace:*"
-  },
-  "dependencies": {
-    "ioredis": "5.10.1"
-  },
-  "devDependencies": {
-    "@zeltjs/core": "workspace:*",
-    "@zeltjs/kv": "workspace:*",
-    "@types/node": "22.19.17"
-  }
-}
