--- packages/kv-driver-redis/src/index.ts
@@ -1,6 +1,0 @@
-export { RedisKVConfig, RedisKVConfig as RedisConfig } from './redis-kv.config';
-export {
-  RedisKVDriver,
-  RedisKVDriver as RedisKVService,
-  RedisKVDriver as RedisKV,
-} from './redis-kv.driver';
