--- packages/kv-driver-redis/src/redis-kv.driver.ts
@@ -1,27 +1,0 @@
-import type { Lifecycle } from '@zeltjs/core';
-import { Injectable, inject, LifecycleManager } from '@zeltjs/core';
-import type { AtomicKVAdaptor, AtomicKVStore, NonEmptyString } from '@zeltjs/kv';
-
-import { RedisKVConfig } from './redis-kv.config';
-import { RedisKVStore } from './redis-kv-store';
-import { ZeltRedis } from './zelt-redis';
-
-@Injectable()
-export class RedisKVDriver implements AtomicKVAdaptor, Lifecycle {
-  private readonly client: ZeltRedis;
-
-  constructor(config = inject(RedisKVConfig), lifecycle = inject(LifecycleManager)) {
-    this.client = new ZeltRedis(config.url, config.options);
-    lifecycle.register(this);
-  }
-
-  namespace<const S extends string>(prefix: NonEmptyString<S>): AtomicKVStore {
-    return new RedisKVStore(this.client, prefix);
-  }
-
-  async startup(): Promise<void> {}
-
-  async shutdown(): Promise<void> {
-    this.client.disconnect();
-  }
-}
