--- packages/kv-driver-redis/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "@9wick/eslint-plugin-strict-type-rules/tsconfig/strictest.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "tsBuildInfoFile": "./dist/.tsbuildinfo",
-    "erasableSyntaxOnly": false,
-    "types": ["node"]
-  },
-  "references": [{ "path": "../core" }, { "path": "../kv" }],
-  "include": ["src/**/*"]
-}
