--- packages/kv-driver-redis/tsdown.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'tsdown';
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm'],
-  dts: true,
-  clean: true,
-  fixedExtension: false,
-  deps: {
-    neverBundle: ['@zeltjs/core', /^@zeltjs\/core\//, '@zeltjs/kv', /^@zeltjs\/kv\//, 'ioredis'],
-  },
-});
