--- packages/kv-driver-redis/vitest.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig, mergeConfig } from 'vitest/config';
-import { sharedConfig } from '../../vitest.shared';
-
-export default mergeConfig(
-  sharedConfig,
-  defineConfig({
-    test: {
-      name: '@zeltjs/kv-driver-redis',
-      include: ['src/**/*.test.ts'],
-    },
-  }),
-);
